restructured test files; cleanups
